% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.Orchestrator.R
\name{Environment.Orchestrator}
\alias{Environment.Orchestrator}
\title{Manage Environment Variables}
\value{
A \code{list} of utility functions:
\itemize{
\item \code{open.config.file()}
\item \code{get.env.variable(name)}
\item \code{cache.env.variable(name, value)}
}
}
\description{
\code{.Renviron} configuration file is well suited to management sensitive information.
This \code{Environment} package makes defining and reading environment variables straightforward.

\code{Environment.Orchestrator()} returns a set of utility functions that can be used to interact with \code{.Renviron} files:
\itemize{
\item The first function: \code{open.config.file}, will open an existing or a new empty \code{.Renviron} configuration file.
Users can then add environment variables by defining \code{'key'='value'} pairs.
\item The second function: \code{get.env.variable}, will read the value of the matching key (name) from the configuration file.
\item The third function: \code{cache.env.variable}, will cache the value of the matching key (name) in the current R session.
}
}
