% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{Environment}
\alias{Environment}
\title{Manage Environment Variables}
\value{
A \code{list} of utility functions:
\itemize{
\item \code{Open.Config.File()}
\item \code{Get.Env.Variable(name)}
\item \code{Cache.Env.Variable(name, value)}
}
}
\description{
\code{.Renviron} configuration file is well suited to management sensitive information.
This \code{Environment} package makes defining and reading environment variables straightforward.

\code{Environment()} returns a set of utility functions that can be used to interact with \code{.Renviron} files:
\itemize{
\item The first function: \code{Open.Config.File}, will open an existing or a new empty \code{.Renviron} configuration file.
Users can then add environment variables by defining \code{'key'='value'} pairs.
\item The second function: \code{Get.Env.Variable}, will read the value of the matching key (name) from the configuration file.
\item The third function: \code{Cache.Env.Variable}, will cache the value of the matching key (name) in the current R session.
}
}
