% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.Exceptions.R
\name{Utility.Exceptions}
\alias{Utility.Exceptions}
\title{Manage Environment Variables}
\value{
A \code{list} of exceptions:
\itemize{
\item \code{NameIsNull()}
\item \code{ValueIsEmpty()}
}
}
\description{
\code{.Renviron} configuration file is well suited to management sensitive information.
This \code{Environment} package makes defining and reading environment variables straightforward.

\code{Utility.Exceptions()} is an internal function that returns a set of exceptions used by \code{Utility.Validation.Service}.
\itemize{
\item The first exception: \code{NameIsNull}, indicates to the user that no environment variable name has been provided.
\item The second exception: \code{ValueIsEmpty}, indicates to the user than no matching value for key (name) could be found in \code{.Renviron} file.
}
}
