% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.Service.R
\name{Utility.Service}
\alias{Utility.Service}
\title{Manage Environment Variables}
\value{
A \code{list} of utility functions:
\itemize{
\item \code{Open.Configuration.File()}
\item \code{GetVariableByName(name)}
}
}
\description{
\code{.Renviron} configuration file is well suited to management sensitive information.
This \code{Environment} package makes defining and reading environment variables straightforward.

\code{Utility.Service()} returns a set of utility functions that can be used to interact with \code{.Renviron} files:
\itemize{
\item The first function: \code{Open.Configuration.File}, will open an existing or a new empty \code{.Renviron} configuration file.
Users can then add environment variables by defining \code{'key'='value'} pairs.
\item The second function: \code{GetVariableByName}, will read the value of the matching key (name) from the configuration file.
}
}
