% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.Validation.R
\name{Utility.Validation}
\alias{Utility.Validation}
\title{Manage Environment Variables}
\value{
A \code{list} of validators:
\itemize{
\item \code{Name()}
\item \code{Value()}
}
}
\description{
\code{.Renviron} configuration file is well suited to management sensitive information.
This \code{Environment} package makes defining and reading environment variables straightforward.

\code{Utility.Validation()} is an internal function that returns a set of validators used by \code{Utility.Service()}.
\itemize{
\item The first validator: \code{Name}, is used to check if a name argument was provided.
\item The second validator: \code{Value}, is used to check if a value for environment variable was found.
}
}
